Renamed sub-components
