+++ src/views/ApartmentApp/components/StudentApplication/components/ApplicantList/components/ApplicantListItem/index.js
