+++ src/views/ApartmentApp/components/StudentApplication/components/HallSelection/components/HallListItem/index.js
