+++ src/views/ApartmentApp/components/StudentApplication/components/HallSelection/index.js
@@ -3,7 +3,7 @@ import 'date-fns';
 import { Grid, Card, CardHeader, CardContent, List, Typography, Button } from '@material-ui/core';
 import AddIcon from '@material-ui/icons/Add';
 import HallListItem from './components/HallListItem';
-import SaveButton from '../ApartAppSaveButton';
+import SaveButton from '../SaveButton';
 import goStalk from '../../../../../../services/goStalk';
 import housing from '../../../../../../services/housing';
 import '../../../../apartmentApp.css';
